Function lib3mf_sys::lib3mf_keystore_addresourcedata

pub unsafe extern "C" fn lib3mf_keystore_addresourcedata(
    pKeyStore: Lib3MF_KeyStore,
    pResourceDataGroup: Lib3MF_ResourceDataGroup,
    pPartPath: Lib3MF_PackagePart,
    eAlgorithm: eLib3MFEncryptionAlgorithm,
    eCompression: eLib3MFCompression,
    nAdditionalAuthenticationDataBufferSize: Lib3MF_uint64,
    pAdditionalAuthenticationDataBuffer: *const Lib3MF_uint8,
    pResourceData: *mut Lib3MF_ResourceData
) -> Lib3MFResult

Add resourcedata to resourcedatagroup element

@param[in] pKeyStore - KeyStore instance. @param[in] pResourceDataGroup - The resource data group where to add this resource data @param[in] pPartPath - The path of the part to be encrypted @param[in] eAlgorithm - The encryption algorithm to be used to encrypt this resource @param[in] eCompression - Whether compression should be used prior to encryption @param[in] nAdditionalAuthenticationDataBufferSize - Number of elements in buffer @param[in] pAdditionalAuthenticationDataBuffer - uint8 buffer of Additional data to be encrypted along the contents for better security @param[out] pResourceData - The data resource instance @return error code or 0 (success)