Type Definition lib3mf_sys::Lib3MFKeyWrappingCallback

pub type Lib3MFKeyWrappingCallback = Option<unsafe extern "C" fn(arg1: Lib3MF_AccessRight, arg2: Lib3MF_uint64, arg3: *const Lib3MF_uint8, arg4: Lib3MF_uint64, arg5: *mut Lib3MF_uint64, arg6: *mut Lib3MF_uint8, arg7: Lib3MF_pvoid, arg8: *mut Lib3MF_uint64)>;

Lib3MFKeyWrappingCallback - A callback used to wrap (encrypt) the content key available in keystore resource group

@param[in] pKEKParams - The information about the parameters used used to wrap the key to the contents @param[in] nInBufferBufferSize - Number of elements in buffer @param[in] pInBufferBuffer - uint8 buffer of Buffer to the input value. When encrypting, this should be the plain key. When decrypting, this should be the key cipher. @param[in] nOutBufferBufferSize - Number of elements in buffer @param[out] pOutBufferNeededCount - will be filled with the count of the written elements, or needed buffer size. @param[out] pOutBufferBuffer - uint8 buffer of Buffer where the data will be placed. When encrypting, this will be the key cipher. When decrypting, this will be the plain key. When buffer is null, neededBytes contains the required bytes to run. @param[in] pUserData - Userdata that is passed to the callback function @param[out] pStatus - The needed/encrypted/decrypted bytes when succeed or zero when error.